Add user to user messaging
